+++ packages/terra-action-footer/src/terra-dev-site/doc/action-footer/Standard.a.doc.jsx
@@ -3,10 +3,9 @@ import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import ActionFooterSrc from '!raw-loader!../../../../src/ActionFooter';
-
 // Example Files
+import ActionFooterPropsTable from '!terra-props-table-loader!../../../../src/ActionFooter';
+import ActionFooterSrc from '!raw-loader!../../../../src/ActionFooter';
 import EmptyExample from '../example/standard/Empty';
 import EmptyExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/standard/Empty';
 import SingleStartActionExample from '../example/standard/SingleStartAction';
@@ -48,6 +47,7 @@ const DocPage = () => (
       {
         componentName: 'Action Footer',
         componentSrc: ActionFooterSrc,
+        componentProps: ActionFooterPropsTable,
       },
     ]}
   />