+++ packages/terra-card/src/terra-dev-site/doc/card/Card.doc.jsx
@@ -3,11 +3,11 @@ import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import CardSrc from '!raw-loader!../../../../src/Card';
-import CardBodySrc from '!raw-loader!../../../../src/CardBody';
-
 // Example Files
+import CardPropsTable from '!terra-props-table-loader!../../../../src/Card';
+import CardSrc from '!raw-loader!../../../../src/Card';
+import CardBodyPropsTable from '!terra-props-table-loader!../../../../src/CardBody';
+import CardBodySrc from '!raw-loader!../../../../src/CardBody';
 import CardDefault from '../example/CardDefault';
 import CardDefaultSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/CardDefault';
 import CardRaised from '../example/CardRaised';
@@ -78,10 +78,12 @@ const DocPage = () => (
       {
         componentName: 'Card',
         componentSrc: CardSrc,
+        componentProps: CardPropsTable,
       },
       {
         componentName: 'Card Body',
         componentSrc: CardBodySrc,
+        componentProps: CardBodyPropsTable,
       },
     ]}
   />