+++ packages/terra-divider/src/terra-dev-site/doc/divider/Divider.1.doc.jsx
@@ -3,10 +3,9 @@ import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import DividerSrc from '!raw-loader!../../../../src/Divider';
-
 // Example Files
+import DividerPropsTable from '!terra-props-table-loader!../../../../src/Divider';
+import DividerSrc from '!raw-loader!../../../../src/Divider';
 import DividerExample from '../example/DividerExample';
 import DividerExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/DividerExample';
 import DividerWithText from '../example/DividerWithText';
@@ -41,6 +40,7 @@ const DocPage = () => (
       {
         componentName: 'Divider',
         componentSrc: DividerSrc,
+        componentProps: DividerPropsTable,
       },
     ]}
   />