+++ packages/terra-doc-template/src/terra-dev-site/doc/doc-template/Placeholder.2.doc.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import DocTemplate from '../../../DocTemplate';
 import ReadMe from '../../../../docs/PLACEHOLDER.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import PlaceholderSrc from '!raw-loader!../../../../src/Placeholder.jsx';
-
 // Example Files
+import PlaceholderPropsTable from '!terra-props-table-loader!../../../../src/Placeholder';
+import PlaceholderSrc from '!raw-loader!../../../../src/Placeholder';
 import PlaceholderDefault from '../example/PlaceholderDefault';
 import PlaceholderDefaultSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/PlaceholderDefault';
 import PlaceholderLight from '../example/PlaceholderLight';
@@ -32,8 +32,9 @@ const DocPage = () => (
     ]}
     propsTables={[
       {
-        componentName: 'PlaceholderSrc',
+        componentName: 'Placeholder',
         componentSrc: PlaceholderSrc,
+        componentProps: PlaceholderPropsTable,
       },
     ]}
   />