+++ packages/terra-form-fieldset/src/terra-dev-site/doc/form-fieldset/FormFieldset.1.doc.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import FieldsetSrc from '!raw-loader!../../../../src/Fieldset';
-
 // Example Files
+import FormFieldsetPropsTable from '!terra-props-table-loader!../../../../src/Fieldset';
+import FormFieldsetSrc from '!raw-loader!../../../../src/Fieldset';
 import FieldsetExamples from '../example/FieldsetExamples';
 
 const DocPage = () => (
@@ -22,8 +22,9 @@ const DocPage = () => (
     ]}
     propsTables={[
       {
-        componentName: 'Field Set',
-        componentSrc: FieldsetSrc,
+        componentName: 'Form Fieldset',
+        componentSrc: FormFieldsetSrc,
+        componentProps: FormFieldsetPropsTable,
       },
     ]}
   />