+++ packages/terra-form-select/src/terra-dev-site/doc/form-select/SearchSelect.11.doc.jsx
@@ -3,12 +3,13 @@ import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import OptionSrc from '!raw-loader!../../../../src/shared/_Option';
-import OptGroupSrc from '!raw-loader!../../../../src/shared/_OptGroup';
-import SearchSelectSrc from '!raw-loader!../../../../src/SearchSelect';
-
 // Example Files
+import SearchSelectPropsTable from '!terra-props-table-loader!../../../../src/SearchSelect';
+import SearchSelectSrc from '!raw-loader!../../../../src/SearchSelect';
+import OptionPropsTable from '!terra-props-table-loader!../../../../src/shared/_Option';
+import OptionSrc from '!raw-loader!../../../../src/shared/_Option';
+import OptGroupPropsTable from '!terra-props-table-loader!../../../../src/shared/_OptGroup';
+import OptGroupSrc from '!raw-loader!../../../../src/shared/_OptGroup';
 import AllowClearExample from '../example/search/AllowClear';
 import AllowClearExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/search/AllowClear';
 import ControlledExample from '../example/search/Controlled';
@@ -30,16 +31,23 @@ const Index = () => (
     packageVersion={version}
     readme={ReadMe}
     srcPath="https://github.com/cerner/terra-core/tree/master/packages/terra-form-select"
-    propsTables={[{
-      componentName: 'SearchSelect',
-      componentSrc: SearchSelectSrc,
-    }, {
-      componentName: 'Option',
-      componentSrc: OptionSrc,
-    }, {
-      componentName: 'OptGroup',
-      componentSrc: OptGroupSrc,
-    }]}
+    propsTables={[
+      {
+        componentName: 'Search Select',
+        componentSrc: SearchSelectSrc,
+        componentProps: SearchSelectPropsTable,
+      },
+      {
+        componentName: 'Option',
+        componentSrc: OptionSrc,
+        componentProps: OptionPropsTable,
+      },
+      {
+        componentName: 'OptGroup',
+        componentSrc: OptGroupSrc,
+        componentProps: OptGroupPropsTable,
+      },
+    ]}
     examples={[{
       title: 'Search',
       description: 'A Search Select allows selecting a single option.',