+++ packages/terra-form-select/src/terra-dev-site/doc/form-select/SingleSelectField.04.doc.jsx
@@ -2,12 +2,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/SingleSelectField.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import SingleSelectFieldSrc from '!raw-loader!../../../../src/SingleSelectField';
-
 // Example Files
+import SingleSelectFieldPropsTable from '!terra-props-table-loader!../../../../src/SingleSelectField';
+import SingleSelectFieldSrc from '!raw-loader!../../../../src/SingleSelectField';
 import SelectFieldExample from '../example/single/SingleSelectField';
 import SelectFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/single/SingleSelectField';
 import RequiredSingleSelectFieldExample from '../example/single/RequiredSingleSelectField';
@@ -19,10 +19,13 @@ const DocPage = () => (
     packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
-    propsTables={[{
-      componentName: 'Single Select Field',
-      componentSrc: SingleSelectFieldSrc,
-    }]}
+    propsTables={[
+      {
+        componentName: 'Single Select Field',
+        componentSrc: SingleSelectFieldSrc,
+        componentProps: SingleSelectFieldPropsTable,
+      },
+    ]}
     examples={[{
       title: 'Single Select Field Example',
       example: <SelectFieldExample />,