+++ packages/terra-signature/src/terra-dev-site/doc/signature/Signature.1.doc.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import SignatureSrc from '!raw-loader!../../../../src/Signature';
-
 // Example Files
+import SignaturePropsTable from '!terra-props-table-loader!../../../../src/Signature';
+import SignatureSrc from '!raw-loader!../../../../src/Signature';
 import SignatureExample from '../example/SignatureExample';
 import SignatureExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/SignatureExample';
 
@@ -27,6 +27,7 @@ const DocPage = () => (
       {
         componentName: 'Signature',
         componentSrc: SignatureSrc,
+        componentProps: SignaturePropsTable,
       },
     ]}
   />