+++ packages/terra-base/CHANGELOG.md
@@ -1,8 +1,11 @@
-Changelog
+ChangeLog
 =========
 
 Unreleased
 ----------
+### Added
+* Props table markdown file import added to doc site page
+
 
 5.17.0 - (August 21, 2019)
 ------------------