+++ packages/terra-base/CHANGELOG.md
@@ -1,8 +1,13 @@
-Changelog
+ChangeLog
 =========
 
 Unreleased
 ----------
+
+5.18.0 - (September 6, 2019)
+------------------
+### Added
+* Props table markdown file import added to doc site page
 
 5.17.0 - (August 21, 2019)
 ------------------